database_query_test/requirements.txt
psycopg2-binary==2.9.9
quixstreams~=2.8.0